+++ src/config.ts
@@ -1,11 +1,3 @@
-let cssRootFolder: string = '/public/styles/'
-
 export default class Config {
-  static get cssRootFolder() {
-    return cssRootFolder
-  }
-
-  static set cssRootFolder(dir: string) {
-    cssRootFolder = dir
-  }
+  static cssRootFolder: string = '/public/styles/';
 }